examples/customer_env_k8s/sample_agent/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
opentelemetry-sdk>=1.24.0
opentelemetry-exporter-otlp-proto-grpc>=1.24.0
opentelemetry-instrumentation-fastapi>=0.45b0
